requirements.txt
fastapi>=0.115.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic-settings>=2.6.0,<3.0.0
sqlalchemy>=2.0,<3
geoalchemy2>=0.15.0,<1.0.0
shapely>=2.0.0,<3.0.0
asyncpg>=0.30.0,<1.0.0
psycopg2-binary>=2.9.0,<3.0.0
alembic>=1.14.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt<4.0.0
argon2-cffi>=23.1.0,<24.0.0
email-validator>=2.0.0,<3.0.0
openai>=1.50.0,<2.0.0
httpx==0.28.1